Reworked delete note function
Now with confirm window.

+++ src/frontend/www/js/script.js
@@ -204,19 +204,25 @@ function updateListName(){
     }
 }
 
-function deleteNoteFunctionalty(){
+function deleteNoteFunctionality(){
 
     // Takes the stored id and parses it correctly
     let LocalStorageid = localStorage.getItem("id");
     let id = parseInt(LocalStorageid);
 
-    // On click - deletes entry in database
-    $("#delete-note-button").click(function () {
+    // Show confirmation window 
+    let confirmWindow = confirm("Är du säker?");
+
+    // If user clicks ok - deletes entry in database
+    if (confirmWindow){
 
         // REST-call
         delete_note(id);
-        
-    })
+
+    // If user clicks cancel show an alert  
+    } else {
+        alert("Avbröt borttagning");
+    }
 }
 
 function deleteListFunctionality () {
@@ -244,5 +250,4 @@ function deleteListFunctionality () {
 displayLists();
 displayNotes();
 showListsInCreateNote();
-deleteNoteFunctionalty();
 addList();
